components: adding new navbar components and use icons
