+++ components/ui/index.js
@@ -3,11 +3,13 @@ import Alert from './Alert'
 import Logo from './Logo'
 import Layout from './Layout'
 import NavBar from './NavBar'
+import NavBarMenu from './NavBarMenu'
 
 export {
 	Logo,
 	Alert,
 	Button,
 	NavBar,
-	Layout
+	Layout,
+	NavBarMenu,
 }
